Rebase errything. Gonna keep my favorite commit message.
After a lot of work I have finally discovered the secret to life. The secret is monads. the secret is inside all things the secret is the art of monadic composition within which we can become one with the great rob rixian enlightenment bestowed by the justin spahr summers for his son josh abernathy died for our sins after he was kastiglioned out of heaven by the objective-father